+++ docs/source/release-notes.rst
@@ -3,10 +3,16 @@ Release Notes
 
 Version 0.24.2
 ______________
-- Plugin:
-  
+- Engine:
+
+  - Support for MockFile.read() on apps testing hooks
+
+- Plugins:
+
   - dataframes (beta):
+
     - Fix: construction from series compatible with pandas
+
     - Fix: clean index name on new dataframe creation